Consolidate SCALE-MS Type identification.
* Create `identifiers` and `_types` modules.
* Migrate some definitions from `scalems.serialization`
  to `scalems.identifiers`
* Replace `scalems.workflow.ResourceType` with
  `scalems.identifiers.TypeIdentifier`

The semantics were not the same. Generally, use cases for `ResourceType`
 were updated to rely on a "type" Identifier rather than a class
 definition, as was implied by `ResourceType`.
